removemono -> remove_mono removeMono

+++ Engines/Wine/Verbs/Remove Mono/script.js
@@ -6,24 +6,27 @@ include(["engines", "wine", "plugins", "regedit"]);
 * Verb to remove mono
 * @returns {Wine} Wine object
 */
-Wine.prototype.removemono = function () {
-    this.uninstall("Mono");
-
-    this.wizard().wait(tr("Please wait..."));
-    this.regedit().deleteKey("HKLM\\Software\\Microsoft\\.NETFramework\\v2.0.50727\\SBSDisabled");
-
-    this.wizard().wait(tr("Please wait..."));
-    this.regedit().deleteKey("HKLM\\Software\\Microsoft\\NET Framework Setup\\NDP\\v3.5");
-
-    this.wizard().wait(tr("Please wait..."));
-    this.regedit().deleteKey("HKLM\\Software\\Microsoft\\NET Framework Setup\\NDP\\v4");
-
-    remove(this.system32directory() + "/mscoree.dll");
-    if (this.architecture() == "amd64")
+Wine.prototype.removeMono = function () {
+    if(this.uninstall("Mono"))
     {
-        remove(this.system64directory() + "/mscoree.dll");
-    }
+        this.wizard().wait(tr("Please wait..."));
+        this.regedit().deleteKey("HKLM\\Software\\Microsoft\\.NETFramework\\v2.0.50727\\SBSDisabled");
 
+        this.wizard().wait(tr("Please wait..."));
+        this.regedit().deleteKey("HKLM\\Software\\Microsoft\\NET Framework Setup\\NDP\\v3.5");
+
+        this.wizard().wait(tr("Please wait..."));
+        this.regedit().deleteKey("HKLM\\Software\\Microsoft\\NET Framework Setup\\NDP\\v4");
+
+        remove(this.system32directory() + "/mscoree.dll");
+        if (this.architecture() == "amd64")
+        {
+            remove(this.system64directory() + "/mscoree.dll");
+        }
+    }
+    else
+        print("COUCOU");
+    
     return this;
 };
 
@@ -36,7 +39,7 @@ var verbImplementation = {
         wine.prefix(container);
         var wizard = SetupWizard(InstallationType.VERBS, "removemono", java.util.Optional.empty());
         wine.wizard(wizard);
-        wine.removemono();
+        wine.removeMono();
         wizard.close();
     }
 };
